+++ loops/times_do.h
@@ -1,6 +1,6 @@
 #ifndef TIMES_DO_H
 #define TIMES_DO_H
 
-void startTimesDo(int iter);
+void startTimesDo(long long int iter);
 
 #endif
